renamed HorizontalLines to GroupRow
